requirements.txt
Django
Pillow
argparse
django-stronghold==0.2.7
djangorestframework==3.1.2
gunicorn==19.3.0
psycopg2==2.6
python-memcached==1.54
six==1.9.0
wheel==0.24.0
wsgiref==0.1.2
git+https://github.com/DheerendraRathor/iitb_oauth_py.git
django-simple-history==1.7.0
django-wysiwyg-redactor==0.4.9
html5lib==0.9999999